requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.2
torch>=1.11.0
gym>=0.21.0
stable-baselines3>=1.6.0
scikit-learn>=1.0.0
scikit-optimize>=0.9.0
pyyaml>=6.0
pytest>=7.0.0
jupyter>=1.0.0
tqdm>=4.64.0 